--- a/writer.go
+++ b/writer.go
@@ -543,13 +543,13 @@ func (p *MediaPlaylist) Encode() *bytes.Buffer {
 		}
 		p.buf.WriteRune(',')
 		p.buf.WriteString(seg.Title)
-		p.buf.WriteString("\n")
+		p.buf.WriteRune('\n')
 		p.buf.WriteString(seg.URI)
 		if p.Args != "" {
 			p.buf.WriteRune('?')
 			p.buf.WriteString(p.Args)
 		}
-		p.buf.WriteString("\n")
+		p.buf.WriteRune('\n')
 	}
 	if p.Closed {
 		p.buf.WriteString("#EXT-X-ENDLIST\n")