Alternative implementation of the improvements by @hori-ryota - all credits for actual changes is theirs. 
This PR cherry picks the biggest impact but lowest risk and maintains ABI compatibility.

This is only an example PR, and will need verification, approval and rebasing before merging.

This branch:

Before:
```
BenchmarkDecodeMasterPlaylist-4   100000             22425 ns/op           12195 B/op         77 allocs/op
BenchmarkDecodeMediaPlaylist-4        50          23947209 ns/op        12279896 B/op     160059 allocs/op
BenchmarkEncodeMasterPlaylist-4  1000000              1023 ns/op              40 B/op         11 allocs/op
BenchmarkEncodeMediaPlaylist-4       100          12931950 ns/op         1748991 B/op      81607 allocs/op
```

After:
```
BenchmarkDecodeMasterPlaylist-4   100000             22534 ns/op           12194 B/op         77 allocs/op
BenchmarkDecodeMediaPlaylist-4       100          16132527 ns/op        10999880 B/op     120059 allocs/op
BenchmarkEncodeMasterPlaylist-4  1000000              1016 ns/op              40 B/op         11 allocs/op
BenchmarkEncodeMediaPlaylist-4       300           4471880 ns/op           45421 B/op        405 allocs/op
```

Benchstat Summary:
```
name                    old time/op    new time/op    delta
DecodeMasterPlaylist-4    22.4µs ± 0%    22.5µs ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-4     23.9ms ± 0%    16.1ms ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-4    1.02µs ± 0%    1.02µs ± 0%   ~             (p=1.000 n=1+1)
EncodeMediaPlaylist-4     12.9ms ± 0%     4.5ms ± 0%   ~             (p=1.000 n=1+1)

name                    old alloc/op   new alloc/op   delta
DecodeMasterPlaylist-4    12.2kB ± 0%    12.2kB ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-4     12.3MB ± 0%    11.0MB ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-4     40.0B ± 0%     40.0B ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-4     1.75MB ± 0%    0.05MB ± 0%   ~             (p=1.000 n=1+1)

name                    old allocs/op  new allocs/op  delta
DecodeMasterPlaylist-4      77.0 ± 0%      77.0 ± 0%   ~     (all samples are equal)
DecodeMediaPlaylist-4       160k ± 0%      120k ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-4      11.0 ± 0%      11.0 ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-4      81.6k ± 0%      0.4k ± 0%   ~             (p=1.000 n=1+1)
```

https://github.com/hori-ryota/m3u8 - feature/improve-perform-decodeplaylist

Before:
```
BenchmarkDecodeMasterPlaylist-4   100000             23203 ns/op           14955 B/op         77 allocs/op
BenchmarkDecodeMediaPlaylist-4        50          24187010 ns/op        12279952 B/op     160060 allocs/op
BenchmarkEncodeMediaPlaylist-4       100          12919057 ns/op         1748992 B/op      81607 allocs/op
```

After:
```
BenchmarkDecodeMasterPlaylist-4   200000             10104 ns/op           12240 B/op         35 allocs/op
BenchmarkDecodeMediaPlaylist-4       100          16193330 ns/op         9200522 B/op      80019 allocs/op
BenchmarkEncodeMediaPlaylist-4       500           3697907 ns/op           21128 B/op        163 allocs/op
```

Benchstat Summary:
```
name                    old time/op    new time/op    delta
DecodeMasterPlaylist-4    23.2µs ± 0%    10.1µs ± 0%   ~     (p=1.000 n=1+1)
DecodeMediaPlaylist-4     24.2ms ± 0%    16.2ms ± 0%   ~     (p=1.000 n=1+1)
EncodeMediaPlaylist-4     12.9ms ± 0%     3.7ms ± 0%   ~     (p=1.000 n=1+1)

name                    old alloc/op   new alloc/op   delta
DecodeMasterPlaylist-4    15.0kB ± 0%    12.2kB ± 0%   ~     (p=1.000 n=1+1)
DecodeMediaPlaylist-4     12.3MB ± 0%     9.2MB ± 0%   ~     (p=1.000 n=1+1)
EncodeMediaPlaylist-4     1.75MB ± 0%    0.02MB ± 0%   ~     (p=1.000 n=1+1)

name                    old allocs/op  new allocs/op  delta
DecodeMasterPlaylist-4      77.0 ± 0%      35.0 ± 0%   ~     (p=1.000 n=1+1)
DecodeMediaPlaylist-4       160k ± 0%       80k ± 0%   ~     (p=1.000 n=1+1)
EncodeMediaPlaylist-4      81.6k ± 0%      0.2k ± 0%   ~     (p=1.000 n=1+1)
```